// client/src/js/bindings/AudioFileResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { UploaderIdentity } from "./UploaderIdentity";

export type AudioFileResponse = { id: string, uploader: UploaderIdentity, original_filename: string, duration_ms: number, file_size_bytes: number, game: string, created_at: number, };